requirements.txt
# pip freeze > requirements.txt

allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
bcrypt==4.1.2
certifi==2023.7.22
cffi==1.16.0
chainmap==1.0.3
charset-normalizer==3.2.0
combomethod==1.0.12
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.2
fabric==3.2.2
fake==0.8
h11==0.14.0
idna==3.4
iniconfig==2.0.0
invoke==2.2.0
nulltype==2.3.1
options==1.4.10
outcome==1.2.0
packaging==23.1
paramiko==3.4.0
pluggy==1.2.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
wrapt==1.16.0
wsproto==1.2.0
